wip: add time to first query
