+++ configs/eslint-config-compass/index.js
@@ -58,7 +58,7 @@ module.exports = {
       rules: { ...tsRules },
     },
     {
-      files: ['**/*.spec.js', '**/*.spec.jsx', '**/*.spec.ts', '**/*.spec.tsx'],
+      files: ['**/*.spec.js', '**/*.spec.jsx', '**/*.spec.ts', '**/*.spec.tsx', '**/*.test.js'],
       env: { mocha: true },
       extends: [...testConfigurations],
     },
