+++ packages/compass-e2e-tests/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  extends: ['@mongodb-js/eslint-config-compass']
+};
