+++ packages/compass-e2e-tests/helpers/commands/connect-with-connection-string.js
@@ -1,0 +1,14 @@
+const Selectors = require('../selectors');
+
+module.exports = function(app) {
+  return async function connectWithConnectionString(
+    connectionString,
+    timeout = 10000
+  ) {
+    await app.client.setValueVisible(
+      Selectors.ConnectionStringInput,
+      connectionString
+    );
+    await app.client.doConnect(timeout);
+  };
+}
