+++ packages/compass-e2e-tests/helpers/compass.js
@@ -7,6 +7,7 @@ const { promisify } = require('util');
 const { Application } = require('spectron');
 const { rebuild } = require('electron-rebuild');
 const debug = require('debug')('compass-e2e-tests');
+
 const {
   run: packageCompass,
   cleanCompileCache,
@@ -14,7 +15,9 @@ const {
   createPackagedStyles
 } = require('hadron-build/commands/release');
 const Selectors = require('./selectors');
-const { delay } = require('./delay');
+const { createUnlockedKeychain } = require('./keychain');
+const { retryWithBackoff } = require('./retry-with-backoff');
+const { addCommands } = require('./commands');
 
 /**
  * @typedef {Object} ExtendedClient
@@ -37,7 +40,6 @@ const cleanCompileCacheAsync = promisify(cleanCompileCache);
 const createCompileCacheAsync = promisify(createCompileCache);
 const createPackagedStylesAsync = promisify(createPackagedStyles);
 
-const MINUTE = 1000 * 60 * 1;
 
 const COMPASS_PATH = path.dirname(
   require.resolve('mongodb-compass/package.json')
@@ -273,372 +275,6 @@ function getCompassBinPath({ appPath, packagerOptions: { name } }) {
 /**
  * @param {ExtendedApplication} app
  */
-function addCommands(app) {
-  // waitForVisible gives better errors than interacting with a non-existing
-  // element
-  app.client.addCommand(
-    'clickVisible',
-    async function clickVisible(selector, timeout = 1000) {
-      await app.client.waitForVisible(selector, timeout);
-      await app.client.click(selector);
-    }
-  );
-
-  app.client.addCommand(
-    'setValueVisible',
-    async function setValueVisible(selector, value, timeout = 1000) {
-      await app.client.waitForVisible(selector, timeout);
-      await app.client.setValue(selector, value);
-    }
-  );
-
-  app.client.addCommand(
-    'waitForConnectionScreen',
-    async function waitForConnectionScreen() {
-      await app.client.waitUntil(
-        async () => {
-          // Compass starts with two windows (one is loading, another is main)
-          // and then one of them is closed. To make sure we are always checking
-          // against existing window, we "focus" the first existing one every
-          // time we run the check (spectron doesn't do it for you automatically
-          // and will fail when certain methods are called on closed windows)
-          await app.client.windowByIndex(0);
-          return await app.client.waitForVisible(Selectors.ConnectSection);
-        },
-        MINUTE,
-        'Expected connection screen to be visible',
-        50
-      );
-    }
-  );
-
-  app.client.addCommand('closeTourModal', async function () {
-    // Wait a bit in any case if it exists or doesn't just so it has a chance to
-    // render if possible
-    await delay(1000);
-    if (await app.client.isExisting(Selectors.FeatureTourModal)) {
-      await app.client.waitUntil(
-        async () => {
-          return await app.client.isVisible(Selectors.FeatureTourModal);
-        },
-        1000,
-        'Expected feature tour modal to be visible',
-        50
-      );
-      // Wait a bit before clicking so that transition is through
-      await delay(100);
-      await app.client.clickVisible(Selectors.CloseFeatureTourModal);
-      await app.client.waitUntil(
-        async () => {
-          return !(await app.client.isExisting(Selectors.FeatureTourModal));
-        },
-        5000,
-        'Expected feature tour modal to disappear after closing it',
-        50
-      );
-    }
-  });
-
-  app.client.addCommand(
-    'closePrivacySettingsModal',
-    async function closePrivacySettingsModal() {
-      // Wait a bit in any case if it exists or doesn't just so it has a chance to
-      // render if possible
-      await delay(1000);
-      if (await app.client.isExisting(Selectors.PrivacySettingsModal)) {
-        await app.client.waitUntil(
-          async () => {
-            return await app.client.isVisible(Selectors.PrivacySettingsModal);
-          },
-          1000,
-          'Expected privacy settings modal to be visible',
-          50
-        );
-        // Wait a bit before clicking so that transition is through
-        await delay(100);
-        await app.client.clickVisible(Selectors.ClosePrivacySettingsButton);
-        await app.client.waitUntil(
-          async () => {
-            return !(await app.client.isExisting(
-              Selectors.PrivacySettingsModal
-            ));
-          },
-          5000,
-          'Expected privacy settings modal to disappear after closing it',
-          50
-        );
-      }
-    }
-  );
-
-  app.client.addCommand('doConnect', async function doConnect(timeout = 10000) {
-    await app.client.clickVisible(Selectors.ConnectButton);
-    // First meaningful thing on the screen after being connected, good enough
-    // indicator that we are connected to the server
-    await app.client.waitForVisible(Selectors.DatabasesTable, timeout);
-  });
-
-  app.client.addCommand(
-    'connectWithConnectionString',
-    async function connectWithConnectionString(
-      connectionString,
-      timeout = 10000
-    ) {
-      await app.client.setValueVisible(
-        Selectors.ConnectionStringInput,
-        connectionString
-      );
-      await app.client.doConnect(timeout);
-    }
-  );
-
-  app.client.addCommand(
-    'connectWithConnectionForm',
-    async function connectWithConnectionForm(
-      {
-        host,
-        port,
-        srvRecord,
-        username,
-        password,
-        authenticationMechanism,
-        gssapiServiceName,
-        replicaSet,
-        tlsAllowInvalidHostnames,
-        sslValidate,
-        tlsCAFile,
-        tlsCertificateKeyFile,
-        sshTunnelHostname,
-        sshTunnelPort,
-        sshTunnelUsername,
-        sshTunnelPassword,
-        sshTunnelIdentityFile
-      },
-      timeout = 10000
-    ) {
-      if (await app.client.isVisible(Selectors.ShowConnectionFormButton)) {
-        await app.client.click(Selectors.ShowConnectionFormButton);
-      }
-
-      await app.client.clickVisible(Selectors.ConnectionFormHostnameTabButton);
-
-      if (typeof host !== 'undefined') {
-        await app.client.setValue(Selectors.ConnectionFormInputHostname, host);
-      }
-
-      if (typeof port !== 'undefined') {
-        await app.client.setValue(Selectors.ConnectionFormInputPort, port);
-      }
-
-      if (srvRecord === true) {
-        await app.client.clickVisible(Selectors.ConnectionFormInputSrvRecord);
-      }
-
-      const authStrategy =
-        authenticationMechanism === 'GSSAPI'
-          ? 'KERBEROS'
-          : authenticationMechanism === 'PLAIN'
-          ? 'LDAP'
-          : authenticationMechanism === 'MONGODB-X509'
-          ? 'X509'
-          : username || password
-          ? 'MONGODB'
-          : 'NONE';
-
-      await app.client.selectByValue(
-        Selectors.ConnectionFormInputAuthStrategy,
-        authStrategy
-      );
-
-      if (typeof username !== 'undefined') {
-        // TODO: No point in having different `name`s in UI, they are not used for
-        // anything and all those map to `username` in driver options anyway
-        if (
-          await app.client.isVisible(
-            Selectors.ConnectionFormInputKerberosPrincipal
-          )
-        ) {
-          await app.client.setValue(
-            Selectors.ConnectionFormInputKerberosPrincipal,
-            username
-          );
-        } else if (
-          await app.client.isVisible(Selectors.ConnectionFormInputLDAPUsername)
-        ) {
-          await app.client.setValue(
-            Selectors.ConnectionFormInputLDAPUsername,
-            username
-          );
-        } else {
-          await app.client.setValue(
-            Selectors.ConnectionFormInputUsername,
-            username
-          );
-        }
-      }
-
-      if (typeof password !== 'undefined') {
-        // TODO: See above
-        if (
-          await app.client.isVisible(Selectors.ConnectionFormInputLDAPPassword)
-        ) {
-          await app.client.setValue(
-            Selectors.ConnectionFormInputLDAPPassword,
-            password
-          );
-        } else {
-          await app.client.setValue(
-            Selectors.ConnectionFormInputPassword,
-            password
-          );
-        }
-      }
-
-      if (typeof gssapiServiceName !== 'undefined') {
-        await app.client.setValue(
-          '[name="kerberos-service-name"]',
-          gssapiServiceName
-        );
-      }
-
-      await app.client.clickVisible('#More_Options');
-
-      if (typeof replicaSet !== 'undefined') {
-        await app.client.setValue(
-          Selectors.ConnectionFormInputReplicaSet,
-          replicaSet
-        );
-      }
-
-      const sslMethod =
-        tlsAllowInvalidHostnames === true || sslValidate === false
-          ? 'UNVALIDATED'
-          : typeof tlsCAFile !== 'undefined' &&
-            typeof tlsCertificateKeyFile !== 'undefined'
-          ? 'ALL'
-          : typeof tlsCAFile !== 'undefined'
-          ? 'SERVER'
-          : /mongodb.net$/.test(host)
-          ? 'SYSTEMCA'
-          : 'NONE';
-
-      await app.client.selectByValue(
-        Selectors.ConnectionFormInputSSLMethod,
-        sslMethod
-      );
-
-      if (['ALL', 'SERVER'].includes(sslMethod)) {
-        // TODO: Can be implemented after https://github.com/mongodb-js/compass/pull/2380
-        throw new Error("Can't test connections that use SSL");
-      }
-
-      const sshTunnel =
-        typeof sshTunnelPassword !== 'undefined'
-          ? 'USER_PASSWORD'
-          : typeof sshTunnelIdentityFile !== 'undefined'
-          ? 'IDENTITY_FILE'
-          : 'NONE';
-
-      if (sshTunnel === 'IDENTITY_FILE') {
-        // TODO: Can be implemented after https://github.com/mongodb-js/compass/pull/2380
-        throw new Error(
-          "Can't test connections that use identity file authentication for SSH tunnel"
-        );
-      }
-
-      await app.client.selectByValue(
-        Selectors.ConnectionFormInputSSHTunnel,
-        sshTunnel
-      );
-
-      if (typeof sshTunnelHostname !== 'undefined') {
-        await app.client.setValue(
-          '[name="sshTunnelHostname"]',
-          sshTunnelHostname
-        );
-      }
-
-      if (typeof sshTunnelPort !== 'undefined') {
-        await app.client.setValue('[name="sshTunnelPort"]', sshTunnelPort);
-      }
-
-      if (typeof sshTunnelUsername !== 'undefined') {
-        await app.client.setValue(
-          '[name="sshTunnelUsername"]',
-          sshTunnelUsername
-        );
-      }
-
-      if (typeof sshTunnelPassword !== 'undefined') {
-        await app.client.setValue(
-          '[name="sshTunnelPassword"]',
-          sshTunnelPassword
-        );
-      }
-
-      await app.client.doConnect(timeout);
-    }
-  );
-
-  app.client.addCommand('disconnect', async function () {
-    // If we are still connecting, let's try cancelling the connection first
-    if (await app.client.isVisible(Selectors.CancelConnectionButton)) {
-      try {
-        await app.client.clickVisible(Selectors.CancelConnectionButton);
-        await app.client.waitUntil(async () => {
-          return !(await app.client.isExisting(
-            Selectors.ConnectionStatusModalContent
-          ),
-          1000,
-          'Expected connection status modal to disappear after cancelling the connection',
-          50);
-        });
-        return;
-      } catch (e) {
-        // If that failed, the button was probably gone before we managed to
-        // click it. Let's go through the whole disconnecting flow now
-      }
-    }
-    app.webContents.send('app:disconnect');
-    await app.client.waitForVisible(Selectors.ConnectSection, 5000);
-    // Show "new connection" section as if we just opened this screen
-    await app.client.clickVisible(Selectors.SidebarNewConnectionButton);
-    await delay(100);
-  });
-
-  app.client.addCommand(
-    'shellEval',
-    async function (str, parse = false, timeout = 10000) {
-      if (!(await app.client.isVisible(Selectors.ShellContent))) {
-        await app.client.clickVisible(Selectors.ShellExpandButton);
-      }
-      await app.client.clickVisible(Selectors.ShellInput);
-      // Might be marked with a deprecation warning, but can be used
-      // https://github.com/webdriverio/webdriverio/issues/2076
-      await app.client.keys(parse === true ? `JSON.stringify(${str})` : str);
-      await app.client.keys('\uE007');
-      await app.client.waitUntil(
-        async () => {
-          return !(await app.client.isVisible(Selectors.ShellLoader));
-        },
-        timeout,
-        `Expected shell evaluation to finish in ${timeout}ms`,
-        50
-      );
-      await delay(50);
-      const output = await app.client.getText(Selectors.ShellOutput);
-      let result = Array.isArray(output) ? output.pop() : output;
-      if (parse === true) {
-        result = JSON.parse(result.replace(/(^['"]|['"]$)/g, ''));
-      }
-      return result;
-    }
-  );
-}
-
-/**
- * @param {ExtendedApplication} app
- */
 function addDebugger(app) {
   const debugClient = debug.extend('webdriver:client');
   // @ts-expect-error getPrototype is not typed in spectron or webdriver but
@@ -702,8 +338,41 @@ async function savePage(
       'HTMLComplete'
     );
     return true;
-  } catch {
+  }
+  catch(err) {
     return false;
+  }
+}
+
+async function beforeTests() {
+  const keychain = createUnlockedKeychain();
+  keychain.activate();
+  const compass = await startCompass();
+
+  // XXX: This seems to be a bit unstable in GitHub CI on macOS machines, for
+  // that reason we want to do a few retries here (in most other cases this
+  // should pass on first attempt)
+  await retryWithBackoff(async () => {
+    await compass.client.waitForConnectionScreen();
+    await compass.client.closeTourModal();
+    await compass.client.closePrivacySettingsModal();
+  });
+
+  return { keychain, compass };
+}
+
+async function afterTests(keychain, compass) {
+  try {
+    if (compass) {
+      if (process.env.CI) {
+        await capturePage(compass);
+        await savePage(compass);
+      }
+      await compass.stop();
+      compass = null;
+    }
+  } finally {
+    keychain.reset();
   }
 }
 
@@ -719,5 +388,7 @@ module.exports = {
   savePage,
   Selectors,
   COMPASS_PATH,
-  LOG_PATH
+  LOG_PATH,
+  beforeTests,
+  afterTests
 };
